Remove separate file for ObserverState
