+++ wpilibc/src/main/native/include/frc/estimator/KalmanFilterLatencyCompensator.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <map>
-#include <frc/estimator/ObserverState.h>
 
 #include <units/units.h>
 
@@ -9,11 +8,25 @@ namespace frc {
 template <int States, int Inputs, int Outputs, typename KalmanTypeFilter>
 class KalmanFilterLatencyCompensator {
  public:
+  class ObserverState {
+   public:
+    const Eigen::Matrix<double, States, 1> xHat;
+    const Eigen::Matrix<double, States, States> errorCovariances;
+    const Eigen::Matrix<double, Inputs, 1> inputs;
+
+    ObserverState(KalmanTypeFilter<States, Inputs, Outputs> observer,
+                  Eigen::Matrix<double, Inputs, 1> u) {
+      xHat = observer->XHat();
+      errorCovariances = observer->P();
+      inputs = u;
+    }
+  };
+
   void AddObserverState(KalmanTypeFilter<States, Inputs, Outputs> observer,
                         Eigen::Matrix<double, Inputs, 1> u,
                         units::second_t timestamp) {
-    m_pastObserverStates.insert(
-        std::pair{timestamp, ObserverState<States, Inputs, Outputs, KalmanTypeFilter>{observer, u}});
+    m_pastObserverStates.insert(std::pair<units::second_t, ObserverState>{
+        timestamp, ObserverState{observer, u}});
 
     if (m_pastObserverStates.size() > kMaxPastObserverStates) {
       m_pastObserverStates.erase(m_pastObserverStates.begin());
@@ -21,13 +34,14 @@ class KalmanFilterLatencyCompensator {
   }
 
   void ApplyPastMeasurement(KalmanTypeFilter<States, Inputs, Outputs>* observer,
-    units::second_t nominalDt,
-    Eigen::Matrix<double, Outputs, 1> y, units::second_t timestamp) {
-    // TODO
+                            units::second_t nominalDt,
+                            Eigen::Matrix<double, Outputs, 1> y,
+                            units::second_t timestamp){
+      // TODO
   };
 
  private:
   constexpr unsigned int kMaxPastObserverStates = 300;
-  std::map<units::second_t, ObserverState<States, Inputs, Outputs, KalmanTypeFilter>> m_pastObserverStates;
+  std::map<units::second_t, ObserverState> m_pastObserverStates;
 };
 }  // namespace frc
